server: Restructure rich-text types somewhat
