+++ server/src/rich-text/types.ts
@@ -54,3 +54,29 @@ export interface InlineElementInput {
   end: number;
   linkTarget?: string | null;
 }
+
+export interface TableCaptionInput {
+  text: string;
+  inlines?: InlineElementInput[] | null;
+}
+
+export const enum CondictLinkTarget {
+  LANGUAGE = 'language',
+  LEMMA = 'lemma',
+  DEFINITION = 'definition',
+  PART_OF_SPEECH = 'part-of-speech',
+}
+
+export interface CondictLink {
+  type: CondictLinkTarget;
+  id: number;
+}
+
+// The input types are exactly compatible with the output types, but we declare
+// them as separate types so they can be changed in the future if necessary.
+
+export type BlockElementJson = BlockElementInput;
+
+export type InlineElementJson = InlineElementInput;
+
+export type TableCaptionJson = TableCaptionInput;
